backend/requirements.txt
flask
flask-cors
requests
google-generativeai
python-dotenv